% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAfreq_in_fasta.R
\name{AAfreq_in_fasta}
\alias{AAfreq_in_fasta}
\title{AA frequency from a fasta file}
\usage{
AAfreq_in_fasta(
  path = "/home/delt/ZHAW/CRC_TRs/data/nfkappab_proteins_CRC_sp.fasta",
  aa_ignore = c("U", "O", "B", "J", "Z", "X", "*", "-", ".", "+", "other")
)
}
\arguments{
\item{path}{str. Path to the fasta file.}

\item{aa_ignore}{character vector of symbols to ignore in the count.}
}
\value{
A data.frame with AA, it's absolute frequency and it's ratios.
}
\description{
AA frequency from a fasta file
}
