% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAfreq_in_prot.R
\name{AAfreq_in_prot}
\alias{AAfreq_in_prot}
\title{Calculate the Amino Acid Frequency in a given AA-Sequence}
\usage{
AAfreq_in_prot(sp)
}
\arguments{
\item{sp}{a XStringSet objects}
}
\value{
A data.frame with AA, it's absolute frequency and it's ratios.
}
\description{
For AA-frequency in a TR-MSA use \code{\link{AAfreq_in_TR}}.
}
