% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAfreq_in_TR.R
\name{AAfreq_in_TR}
\alias{AAfreq_in_TR}
\title{Calculate the Amino Acid Frequency in a given TR-MSA}
\usage{
AAfreq_in_TR(tr_all_sp, includeSpaces = FALSE)
}
\arguments{
\item{tr_all_sp}{a data.frame with TR results imported through \code{\link{load_tr_annotations}}. Requires to have the MSA included as column.}
}
\value{
A data.frame with AA, it's absolute frequency and it's ratios.
}
\description{
Calculate the Amino Acid Frequency in a given TR-MSA
}
