% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AAratio_vs_Disorderpropensity.R
\name{AAratio_vs_Disorderpropensity}
\alias{AAratio_vs_Disorderpropensity}
\title{Amino Acid Ratio Plotted Against Disorderpropensity}
\usage{
AAratio_vs_Disorderpropensity(tr_all, plot_title = NULL, sp_overall = FALSE)
}
\arguments{
\item{tr_all}{a data.frame with TR results imported through \code{\link{load_tr_annotations}}. Requires to have the MSA included as column.}

\item{plot_title}{(str) Title of the plot}
}
\value{
ggplot object
}
\description{
Amino Acid Ratio Plotted Against Disorderpropensity
}
