% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_prot_sequence.R
\name{download_prot_sequence}
\alias{download_prot_sequence}
\title{Download Protein Sequence}
\usage{
download_prot_sequence(prot_id)
}
\arguments{
\item{prot_id}{str. with the Protein identifier.}
}
\value{
a XStringSet objects
}
\description{
Download Protein Sequence
}
