% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTRbyGene.R
\name{getTRbyGene}
\alias{getTRbyGene}
\title{Filter for Tandem Repeats by Gene Name}
\usage{
getTRbyGene(tr_sp, genes)
}
\arguments{
\item{tr_sp}{a data.frame with TR results imported through \code{\link{load_tr_annotations}} and left-joined meta data from SwissProt through \code{\link{load_swissprot}}.}

\item{genes}{\code{str} of gene names. For details see \url{https://www.uniprot.org/help/gene_name}}
}
\description{
Filter for Tandem Repeats by Gene Name
}
