% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_kinome.R
\name{load_kinome}
\alias{load_kinome}
\title{Load Human Proteome Kinome}
\usage{
load_kinome(url, path, OnlyIDs = TRUE)
}
\arguments{
\item{url}{string of the query to the UniProt REST API to download the protein sequences as fasta file.}

\item{path}{string with the filepath to the .fasta file of the protein sequences}

\item{OnlyIDs}{if TRUE, only the the ProteinIDs are return. If FALSE a XString class object of the AAStringSet is returned.}
}
\description{
Load Human Proteome Kinome
}
\examples{
url_kin <- "https://www.uniprot.org/uniprot/?query=ec:2.7.10.-\%20OR\%20ec:2.7.11.-\%20OR\%20ec:2.7.12.-\%20OR\%20ec:2.7.13.-\%20OR\%20ec:2.7.14.-\%20OR\%20ec:2.7.99.-&format=fasta&sort=score&fil=proteome:UP000005640\%20AND\%20reviewed:yes\%20AND\%20organism:\%22Homo\%20sapiens\%20(Human)\%20[9606]\%22"
load_kinome(url_kin, "/myProject/url_kin.fasta")
}
