% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_swissprot.R
\name{load_swissprot}
\alias{load_swissprot}
\title{Load SwissProt Meta Data}
\usage{
load_swissprot(path, tr_all)
}
\arguments{
\item{path}{Path to .tsv file with UniProt/SwissProt information}

\item{tr_all}{a data.frame with TR results imported through \code{\link{function}}}
}
\description{
Load SwissProt Meta Data
}
