% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_tr_annotations.R
\name{load_tr_annotations}
\alias{load_tr_annotations}
\title{Load Tandem Repeat Annotation Data}
\usage{
load_tr_annotations(path)
}
\arguments{
\item{path}{}
}
\value{
Dataframe object with the results from TRAL and a characterisation for the TR-unit size.
}
\description{
Load Tandem Repeat Annotation Data
}
