% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paperfigure.R
\name{paper.figure}
\alias{paper.figure}
\title{ggplot function to prepare figures for publication}
\usage{
paper.figure(p, x.axis.text.angle = 90, x.axis.text.hjust = NULL)
}
\arguments{
\item{p}{ggplot object}

\item{x.axis.text.angle}{angle of x-axis text}

\item{x.axis.text.hjust}{hjust for adjusting the height of the x-axis text}
}
\value{
ggplot object
}
\description{
ggplot function to prepare figures for publication
}
