% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_AAratio.R
\name{plot_AAratio}
\alias{plot_AAratio}
\title{Plot AA ratio}
\usage{
plot_AAratio(df = AAfreq, plot_title = NULL, ylim = 1)
}
\arguments{
\item{df}{data.frame. AA frequency count obtained through \code{\link{AAfreq_in_TR}}.}

\item{plot_title}{str. Plot title.}

\item{ylim}{int. Scale the y-axis maximum value.}
}
\value{
ggplot object.
}
\description{
Plot AA ratio
}
\examples{
plot_AAratio(df = AAfreq_in_TR(tr_all_sp))
}
