% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protein_id_by_number_of_TR.R
\name{protein_id_by_number_of_TR}
\alias{protein_id_by_number_of_TR}
\title{Get the Protein ID from the Number of TR}
\usage{
protein_id_by_number_of_TR(tr_dataframe, no_TR)
}
\arguments{
\item{tr_dataframe}{a data.frame with TR results imported through \code{\link{function}}}

\item{no_TR}{number of TRs per protein.}
}
\value{
the Protein ID toghether with the number of detected TRs
}
\description{
Get the Protein ID from the Number of TR
}
