% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TR_location.R
\name{TR_location}
\alias{TR_location}
\title{Display TR center location}
\usage{
TR_location(tr_all_sp, byTRtype = FALSE, byProt = TRUE,
  paperfigure = FALSE)
}
\arguments{
\item{tr_all_sp}{a data.frame with TR results imported through \code{\link{load_tr_annotations}}}

\item{byTRtype}{boolean. If TRUE, function returns plot filled by TR-type}

\item{byProt}{boolean. IF TRUE, function returns plot filled by each TR.}

\item{paperfigure}{boolean. If TRUE, adapts figure for publication (larger font size, ...). Modifications can be done in \code{\link{paper.figure}}}
}
\value{
ggplot object
}
\description{
The TR center is normalized to reduce boundary effects. \cr
The center-bias comes from boundary effects from the simple center/Length metric employed; \cr
if the TR covers a significant fraction of the protein, then it's center necessarily falls near the middle.\cr
 We can compensate for this by normalizing over only the valid center locations:\cr
 \deqn{x = \frac{center - l_{eff}*n_{eff}/2}{N-l_{eff}*n_{eff}}} \cr
 \cr TR center locations of 0 and 1 refere to the N- and C-terminus respectively.
}
