% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TRdist_in_proteome.R
\name{TRdist_in_proteome}
\alias{TRdist_in_proteome}
\title{Tandem Repeat Distribution in Proteome}
\usage{
TRdist_in_proteome(tr_all_sp, paperfigure = FALSE, max_pvalue = 0.01,
  max_l_effective = 80, max_n_effective = 40)
}
\arguments{
\item{tr_all_sp}{a data.frame with TR results imported through \code{\link{load_tr_annotations}}}

\item{paperfigure}{boolean. If TRUE, adapts figure for publication (larger font size, ...)}

\item{max_pvalue}{upper threshold of p-value. \code{tr_all_sp} is subsetted for pvalue <= max_pvalue.}

\item{max_l_effective}{upper threshold of l_effective \code{tr_all_sp} is subsetted for l_effective <= max_l_effective}

\item{max_n_effective}{upper threshold of n_effective \code{tr_all_sp} is subsetted for n_effective <= max_n_effective}
}
\value{
Distribution (Heatmap) of tandem repeats (TRs) in Swiss-Prot as a function of their repeat unit length $l_{effective} <= 80$ (x-Axis, x1) and their number of repeat units $n_{effective} <= 40$ (x2, y-Axis). Darker colour indicates a larger number of TRs with a specific length and number of repeats.
}
\description{
Tandem Repeat Distribution in Proteome
}
